+++ dev-packages/cli-service/src/webpack/config/dev-server-config-factory.ts
@@ -1,12 +1,12 @@
 
-import { BACKEND_TARGET, CliContext, getPort } from '@malagu/cli-common';
+import { BACKEND_TARGET, CliContext, ConfigUtil } from '@malagu/cli-common';
 import * as WebpackChain from 'webpack-chain';
 
 export class DevServerConfigFactory {
 
     create(config: WebpackChain, context: CliContext, target: string) {
         const { cfg, port, open } = context;
-        const realPort = getPort(cfg, target, port);
+        const realPort = ConfigUtil.getPort(cfg, target, port);
         config
             .devServer
                 .port(realPort)