+++ dev-packages/cli-service/src/webpack/config/malagu-yaml-config-factory.ts
@@ -1,11 +1,11 @@
 
-import { getWebpackConfig, CliContext } from '@malagu/cli-common';
+import { ConfigUtil, CliContext } from '@malagu/cli-common';
 import * as WebpackChain from 'webpack-chain';
 
 export class MalaguYamlConfigFactory {
     create(config: WebpackChain, context: CliContext, target: string) {
         const { cfg } = context;
-        config.merge(getWebpackConfig(cfg, target).config || {});
+        config.merge(ConfigUtil.getWebpackConfig(cfg, target).config || {});
     }
 
     support(context: CliContext, target: string): boolean {