+++ dev-packages/cli-service/src/webpack/config/output-config-factory.ts
@@ -1,5 +1,5 @@
 
-import { BACKEND_TARGET, CliContext, setProjectHomePath, getProjectHomePathForTarget } from '@malagu/cli-common';
+import { BACKEND_TARGET, CliContext, PathUtil } from '@malagu/cli-common';
 import * as path from 'path';
 import * as WebpackChain from 'webpack-chain';
 
@@ -7,9 +7,9 @@ export class OutputConfigFactory {
     create(config: WebpackChain, context: CliContext, target: string) {
         const { dev, output, runtime } = context;
         if (output) {
-            setProjectHomePath(path.join(process.cwd(), output));
+            PathUtil.setProjectHomePath(path.join(process.cwd(), output));
         }
-        const outputPath = path.join(getProjectHomePathForTarget(target, runtime), 'dist');
+        const outputPath = path.join(PathUtil.getProjectHomePathForTarget(target, runtime), 'dist');
 
         config.output.path(outputPath);