+++ dev-packages/compression/src/webpack-hook.ts
@@ -1,10 +1,10 @@
-import { WebpackContext, ConfigurationContext, FRONTEND_TARGET, getWebpackConfig } from '@malagu/cli-service';
+import { WebpackContext, ConfigurationContext, FRONTEND_TARGET, ConfigUtil } from '@malagu/cli-service';
 
 export default async (context: WebpackContext) => {
     const { configurations, dev, cfg } = context;
     const configuration = ConfigurationContext.getConfiguration(FRONTEND_TARGET, configurations);
     if (configuration) {
-        const pluginConfig = getWebpackConfig(cfg, FRONTEND_TARGET).compressionWebpackPlugin;
+        const pluginConfig = ConfigUtil.getWebpackConfig(cfg, FRONTEND_TARGET).compressionWebpackPlugin;
         if (!pluginConfig.disable && !dev) {
             delete pluginConfig.disable;
             const CompressionPlugin = require('compression-webpack-plugin');